fix fread

+++ R/fun_loadTxtToStack.R
@@ -39,7 +39,7 @@ loadTxtToStack<-function(fn_path=NULL,
   if (!file.exists(paste(fn_path,fn_file,sep='/'))) {stop("could not find file")}
   if (is.null(fn_rasterDBPath)) {stop("Please, specifify a path for storing raster files")}
 
-  rawMatrix<-data.table::fread(paste(fn_path,fn_file,sep='/'),sep='\t',header = T,check.names = F,colClasses = 'numeric')
+  rawMatrix<-data.table::fread(paste(fn_path,fn_file,sep='/'),sep='\t',header = T,check.names = F,colClasses = 'double')
 
 
   rasterMatrix<-lapply(fn_cols,
